Separate the InMemoryKafka modules
