+++ lib/sanbase/api_call_data/in_memory_kafka/in_memory_kafka_producer.ex
@@ -40,19 +40,3 @@ defmodule Sanbase.InMemoryKafka.Producer do
     Agent.get(producer, & &1)
   end
 end
-
-defmodule Sanbase.InMemoryKafka.Supervisor do
-  use Supervisor
-
-  def start_link(opts) do
-    Supervisor.start_link(__MODULE__, opts)
-  end
-
-  def init(opts) do
-    children = [
-      {Sanbase.InMemoryKafka.Producer, opts}
-    ]
-
-    Supervisor.init(children, strategy: :one_for_one)
-  end
-end
